Curso java orientação a objeto formação alura - Desafio final do curso, criação da pagina principal e criando objetos como podcast, musicas para exibir titulo, autor, repodução, curtidas, album e descriçao

+++ src/br/com/alura/DesafioDoCurso/Audio.java
@@ -5,7 +5,7 @@ public class Audio {
     private int duracao ;
     private int totalReproduções;
     private int  totalCurtidas ;
-    private double classificacao;
+    private int classificacao;
 
     public int getDuracao() {
         return duracao;
